# docs/swagger.yaml
openapi: 3.0.0
info:
  title: Socious ID API
  version: 2.2.6
servers:
  - url: https://dev-id.socious.io/
    description: Development server (uses test data)
  - url: https://id.socious.io/
    description: Production server (uses live data)
  - url: http://localhost:3000
    description: Local server (for development purposes)
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token required for user authentication
    CookieAuth:
      type: apiKey
      in: cookie
      name: socious-id-session
      description: Cookie-based authentication for web sessions
    ClientSecret:
      type: apiKey
      in: header
      name: X-Client-Secret
      description: Client secret key required for accessing protected endpoints
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
      required:
        - error
...
paths:
...
  /payments/cards:
    get:
      summary: Get user payment cards
      description: Retrieves a paginated list of payment cards associated with the authenticated user
      tags:
        - Payments
      security:
        - BearerAuth: []
        - CookieAuth: []
      parameters:
        - name: page
          in: query
          required: false
          description: Page number for pagination
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          required: false
          description: Number of items per page
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: List of payment cards retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  page:
                    type: integer
                    description: Current page number
                  limit:
                    type: integer
                    description: Number of items per page
                  results:
                    type: array
                    items:
                      $ref: '#/components/schemas/Card'
                  total:
                    type: integer
                    description: Total number of cards
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized - user not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      summary: Add a new payment card
      description: Adds a new payment card for the authenticated user
      tags:
        - Payments
      security:
        - BearerAuth: []
        - CookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddCardForm'
      responses:
        '200':
          description: Card added successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Card'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized - user not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /payments/cards/{id}:
    get:
      summary: Get payment card by ID
      description: Retrieves details of a specific payment card by its ID
      tags:
        - Payments
      security:
        - BearerAuth: []
        - CookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: UUID of the payment card
          schema:
            type: string
            format: uuid
      responses:
        '302':
          description: Card found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Card'
        '400':
          description: Bad request or card not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized - user not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      summary: Update payment card
      description: Updates a specific payment card by its ID
      tags:
        - Payments
      security:
        - BearerAuth: []
        - CookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: UUID of the payment card
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddCardForm'
      responses:
        '200':
          description: Card updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Card'
        '400':
          description: Bad request or card not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized - user not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      summary: Delete payment card
      description: Deletes a specific payment card by its ID
      tags:
        - Payments
      security:
        - BearerAuth: []
        - CookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: UUID of the payment card
          schema:
            type: string
            format: uuid
      responses:
        '400':
          description: Success or error
          content:
            application/json:
              schema:
                oneOf:
                  - type: object
                    properties:
                      message:
                        type: string
                        example: success
                  - $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized - user not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /payments/wallets:
    get:
      summary: Get user wallets
      description: Retrieves a list of wallets associated with the authenticated user
      tags:
        - Payments
      security:
        - BearerAuth: []
        - CookieAuth: []
      responses:
        '302':
          description: Wallets found
          content:
            application/json:
              schema:
                type: object
                properties:
                  wallets:
                    type: array
                    items:
                      $ref: '#/components/schemas/Wallet'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized - user not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      summary: Add or update a wallet
      description: Adds a new wallet or updates an existing one for the authenticated user
      tags:
        - Payments
      security:
        - BearerAuth: []
        - CookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddWalletForm'
      responses:
        '400':
          description: Success or error
          content:
            application/json:
              schema:
                oneOf:
                  - type: object
                    properties:
                      wallet:
                        $ref: '#/components/schemas/Wallet'
                  - $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized - user not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'